Create non-divisible-subset.cpp
https://www.hackerrank.com/challenges/non-divisible-subset
This solution got 100 scores. However, I think when N = 1, the output should be 0, instead of 1. It violates the condition that the sum of any 2 numbers in S' are not evenly divisible by K. It is only one number in given set S. So the size of maximum subset S' should be 0.